[packaging] Untangle apispecifc metapackages
Previously, apispecific metapackages for newer api levels would
depend on the old ones, to avoid duplicating dependencies between
multiple Package blocks.

This worked great on Q (API 29), but it's pretty inflexible for
the needs of R (API 30).

Let's move the debian/control generation to a specific script, which
allows to combine multiple dependencies around. This MUST be the source
of truth of the metapackages contents - i.e. you should modify this script
and run it to update the debian/control contents.

Apispecific packages are thus now untangled, and a device shouldn't have
packages for different apilevels installed. Every apispecific packages
now specify a Provides field for the given variant.

Signed-off-by: Eugenio Paolantonio (g7) <me@medesimo.eu>