[adaptation-hybris] Decouple adaptation-hybris from phosh
adaptation-hybris depending on phosh (through phosh-config-hwcomposer) is wrong as it
should be a generic package as it's installed in the base gsi recipe.

The dependency of phosh-config-hwcomposer has been removed from adaptation-hybris, and the
package has been renamed to adaptation-hybris-common.

New packages -phosh and -cutie have been added for phosh and cutie respectively. You can
safely install those during image building.

Older installations are grandfathered in through the old adaptation-hybris package, which
is now a transitional package that installs -common and -phosh.

Signed-off-by: Eugenio Paolantonio (g7) <me@medesimo.eu>